+++ src/main/java/ru/yandex/practicum/filmorate/service/FilmService.java
@@ -2,6 +2,7 @@ package ru.yandex.practicum.filmorate.service;
 
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
+import org.springframework.jdbc.core.JdbcTemplate;
 import org.springframework.stereotype.Service;
 import ru.yandex.practicum.filmorate.dal.FilmStorage;
 import ru.yandex.practicum.filmorate.dal.db.storage.ListGenreDbStorage;
@@ -10,11 +11,13 @@ import ru.yandex.practicum.filmorate.exception.EntityNotFoundException;
 import ru.yandex.practicum.filmorate.exception.ValidationException;
 import ru.yandex.practicum.filmorate.model.Film;
 import ru.yandex.practicum.filmorate.model.Genre;
+import ru.yandex.practicum.filmorate.model.Like;
+import ru.yandex.practicum.filmorate.model.Mpa;
 
+import java.sql.ResultSet;
+import java.sql.SQLException;
 import java.time.LocalDate;
-import java.util.Collection;
-import java.util.List;
-import java.util.Optional;
+import java.util.*;
 
 @Service
 @Slf4j
@@ -23,7 +26,7 @@ public class FilmService implements IntefaceService<Film> {
     private static final LocalDate DATE_MARK = LocalDate.of(1895, 12, 28);
     private final FilmStorage filmStorage;
     private final UserService userService;
-    private final BuilderFilm builderFilm;
+    private final BuilderFilm builderFilm = new BuilderFilm(new JdbcTemplate());
     private final GenreService genreService;
     private final ListGenreDbStorage listGenreDbStorage;
     private final MpaService mpaService;
@@ -115,5 +118,4 @@ public class FilmService implements IntefaceService<Film> {
             throw new ValidationException(s);
         }
     }
-
 }
