+++ src/test/java/ru/yandex/practicum/filmorate/FilmorateApplicationTests.java
@@ -1,0 +1,11 @@
+package ru.yandex.practicum.filmorate;
+
+import org.junit.jupiter.api.Test;
+
+class FilmorateApplicationTests {
+
+    @Test
+    void contextLoads() {
+    }
+
+}
